feat: add spaceship skill
z, x skills

+++ src/main/java/org/newdawn/spaceinvaders/entity/ShotEntity.java
@@ -25,7 +25,7 @@ public class ShotEntity extends Entity {
 	 * @param x      The initial x location of the shot
 	 * @param y      The initial y location of the shot
 	 */
-	public ShotEntity(Game game, String sprite, int x, int y, int power, int direction) {
+	public ShotEntity(Game game, String sprite, int x, int y, int power, double direction) {
 
 		super(sprite, x, y);
